requirements.txt
arrow==1.2.3
asgiref==3.6.0
binaryornot==0.4.4
certifi==2022.12.7
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
cookiecutter==1.7.0
crispy-tailwind==0.5.0
dj-database-url==1.3.0
Django==4.2
django-browser-reload==1.7.0
django-crispy-forms==2.0
django-filter==23.1
django-tailwind==3.5.0
docopt==0.6.2
future==0.18.3
gunicorn==20.1.0
idna==3.4
Jinja2==3.1.2
jinja2-time==0.2.0
MarkupSafe==2.1.2
pipreqs==0.4.11
poyo==0.5.0
psycopg2-binary==2.9.6
python-dateutil==2.8.2
python-slugify==8.0.1
PyYAML==6.0
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
text-unidecode==1.3
typing_extensions==4.5.0
urllib3==1.26.15
whichcraft==0.6.1
yarg==0.1.9